+++ docs/source/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Latest
 ------
 
+- Renamed :code:`List.iter()` to :code:`List.items()`.
 - Renamed sections to layers; it reflects their function better.
 
 SunsetSettings 0.7.1 (2026-02-09)